# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AddHistoryRequest": {
        "properties": {
          "beer_id": {
            "title": "Beer Id",
            "type": "string"
          },
          "rating": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RatingValue"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "beer_id"
        ],
        "title": "AddHistoryRequest",
        "type": "object"
      },
      "BeerPayload": {
        "properties": {
          "brewery": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brewery"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "flavor_vector": {
            "items": {
              "type": "number"
            },
            "title": "Flavor Vector",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "style": {
            "title": "Style",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "style",
          "flavor_vector"
        ],
        "title": "BeerPayload",
        "type": "object"
      },
      "ChallengeComparisonResponse": {
        "properties": {
          "challenger_persona": {
            "$ref": "#/components/schemas/PersonaSummary"
          },
          "different": {
            "items": {
              "type": "string"
            },
            "title": "Different",
            "type": "array"
          },
          "friend_persona": {
            "$ref": "#/components/schemas/PersonaSummary"
          },
          "shared": {
            "items": {
              "type": "string"
            },
            "title": "Shared",
            "type": "array"
          },
          "similarity": {
            "title": "Similarity",
            "type": "number"
          }
        },
        "required": [
          "similarity",
          "shared",
          "different",
          "challenger_persona",
          "friend_persona"
        ],
        "title": "ChallengeComparisonResponse",
        "type": "object"
      },
      "ChallengeTokenResponse": {
        "properties": {
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "ChallengeTokenResponse",
        "type": "object"
      },
      "ChallengeVectorRequest": {
        "properties": {
          "vector": {
            "items": {
              "type": "number"
            },
            "title": "Vector",
            "type": "array"
          }
        },
        "required": [
          "vector"
        ],
        "title": "ChallengeVectorRequest",
        "type": "object"
      },
...
      "FlavorVector": {
        "properties": {
          "adventure": {
            "title": "Adventure",
            "type": "number"
          },
          "bitterness": {
            "title": "Bitterness",
            "type": "number"
          },
          "body": {
            "title": "Body",
            "type": "number"
          },
          "fruitiness": {
            "title": "Fruitiness",
            "type": "number"
          },
          "roast": {
            "title": "Roast",
            "type": "number"
          },
          "sourness": {
            "title": "Sourness",
            "type": "number"
          },
          "sweetness": {
            "title": "Sweetness",
            "type": "number"
          }
        },
        "required": [
          "bitterness",
          "sweetness",
          "fruitiness",
          "roast",
          "sourness",
          "body",
          "adventure"
        ],
        "title": "FlavorVector",
        "type": "object"
      },
      "GroupRecommendationResponse": {
        "properties": {
          "group_vector": {
            "items": {
              "type": "number"
            },
            "title": "Group Vector",
            "type": "array"
          },
          "high_variance": {
            "title": "High Variance",
            "type": "boolean"
          }
        },
        "required": [
          "group_vector",
          "high_variance"
        ],
        "title": "GroupRecommendationResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HistoryEntry": {
        "properties": {
          "beer_id": {
            "title": "Beer Id",
            "type": "string"
          },
          "rating": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RatingValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "tried_at": {
            "title": "Tried At",
            "type": "string"
          }
        },
        "required": [
          "beer_id",
          "rating",
          "tried_at"
        ],
        "title": "HistoryEntry",
        "type": "object"
      },
      "HistoryResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/HistoryEntry"
            },
            "title": "Entries",
            "type": "array"
          }
        },
        "required": [
          "entries"
        ],
        "title": "HistoryResponse",
        "type": "object"
      },
...
      "LeaderboardEntryResponse": {
        "properties": {
          "persona_icon": {
            "title": "Persona Icon",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "recommendation_count": {
            "title": "Recommendation Count",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "persona_icon",
          "recommendation_count",
          "rank"
        ],
        "title": "LeaderboardEntryResponse",
        "type": "object"
      },
      "LeaderboardResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntryResponse"
            },
            "title": "Entries",
            "type": "array"
          },
          "viewer_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewer Rank"
          }
        },
        "required": [
          "entries",
          "viewer_rank"
        ],
        "title": "LeaderboardResponse",
        "type": "object"
      },
      "OkResponse": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "type": "object"
      },
      "PersonaResponse": {
        "properties": {
          "persona": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonaSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "persona"
        ],
        "title": "PersonaResponse",
        "type": "object"
      },
      "PersonaSummary": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "icon"
        ],
        "title": "PersonaSummary",
        "type": "object"
      },
      "ProfileResponse": {
        "properties": {
          "user_id": {
            "title": "User Id",
            "type": "string"
          },
          "vector": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector"
          }
        },
        "required": [
          "user_id",
          "vector"
        ],
        "title": "ProfileResponse",
        "type": "object"
      },
      "RateBeerRequest": {
        "properties": {
          "beer": {
            "$ref": "#/components/schemas/RatedBeerInput"
          },
          "rating": {
            "$ref": "#/components/schemas/RatingValue"
          }
        },
        "required": [
          "beer",
          "rating"
        ],
        "title": "RateBeerRequest",
        "type": "object"
      },
      "RatedBeerInput": {
        "properties": {
          "brewery": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brewery"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "flavor_vector": {
            "items": {
              "type": "number"
            },
            "title": "Flavor Vector",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "style": {
            "title": "Style",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "style",
          "flavor_vector"
        ],
        "title": "RatedBeerInput",
        "type": "object"
      },
      "RatingValue": {
        "enum": [
          "loved",
          "fine",
          "disliked"
        ],
        "title": "RatingValue",
        "type": "string"
      },
      "RecommendationRequest": {
        "properties": {
          "beers": {
            "items": {
              "$ref": "#/components/schemas/BeerPayload"
            },
            "title": "Beers",
            "type": "array"
          },
          "taste_vector": {
            "$ref": "#/components/schemas/FlavorVector"
          }
        },
        "required": [
          "taste_vector",
          "beers"
        ],
        "title": "RecommendationRequest",
        "type": "object"
      },
      "RecommendationResponse": {
        "properties": {
          "adventurous": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BeerPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "backup": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BeerPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "best": {
            "$ref": "#/components/schemas/BeerPayload"
          },
          "explanations": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Explanations",
            "type": "object"
          }
        },
        "required": [
          "best",
          "backup",
          "adventurous",
          "explanations"
        ],
        "title": "RecommendationResponse",
        "type": "object"
      },
      "SaveProfileRequest": {
        "properties": {
          "vector": {
            "items": {
              "type": "number"
            },
            "title": "Vector",
            "type": "array"
          }
        },
        "required": [
          "vector"
        ],
        "title": "SaveProfileRequest",
        "type": "object"
      },
...
      "UpdatedVectorResponse": {
        "properties": {
          "updated_vector": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated Vector"
          }
        },
        "required": [
          "updated_vector"
        ],
        "title": "UpdatedVectorResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Beerolog API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/challenges": {
      "post": {
        "operationId": "createChallenge",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Challenge",
        "tags": [
          "challenges"
        ]
      }
    },
    "/challenges/{token}/compare": {
      "post": {
        "operationId": "compareChallenge",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeVectorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare",
        "tags": [
          "challenges"
        ]
      }
    },
...
    "/recommendations/": {
      "post": {
        "operationId": "recommendBeers",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recommend",
        "tags": [
          "recommendations"
        ]
      }
    },
...
    "/sessions/{session_id}/recommend": {
      "get": {
        "operationId": "getGroupRecommendation",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupRecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recommend",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/users/me/history": {
      "get": {
        "operationId": "getMyHistory",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My History",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "addToMyHistory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddHistoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add To My History",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/persona": {
      "get": {
        "operationId": "getMyPersona",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Persona",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/profile": {
      "get": {
        "operationId": "getMyProfile",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get My Profile",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "saveMyProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save My Profile",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/rate": {
      "post": {
        "operationId": "rateMyBeer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateBeerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatedVectorResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rate Beer",
        "tags": [
          "users"
        ]
      }
    },
    "/venues/{venue_id}/leaderboard": {
      "get": {
        "operationId": "getVenueLeaderboard",
        "parameters": [
          {
            "in": "path",
            "name": "venue_id",
            "required": true,
            "schema": {
              "title": "Venue Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Venue Leaderboard",
        "tags": [
          "leaderboard"
        ]
      }
    },
...
  }
}